Relocate cloudprober/rtc package to targets/rtc/rtcservice. rtcservice is relevant only to targets so this improves code structure.
ORIGINAL_AUTHOR=Manu Garg <manugarg@gmail.com>
PiperOrigin-RevId: 164553980
